Rename package "core/promises" -> to "core/promise"
